Add methods to (re/un)schedule expired files cleanup actions.
The action is scheduled to run immediately. After the cleanup happens,
if there are more files pending deletion a new action is scheduled
to run immediately, otherwise a new action is scheduled to run in 24h
(but this can be adjusted with a new filter,
woocommerce_delete_expired_rendered_template_files_interval).

Also add two new debug tools, one to: (re)schedule the action
and another one to unschedule it.