Navigation: Change error_log call to doing_it_wrong
There's a unit test that specifically creates the conditions that
cause this particular error to get triggered. Unfortunately, this
makes the error message get output in the middle of the unit tests,
like this:

```php
 ............................................................. 1952 / 3898 ( 50%)
[16-Dec-2023 00:36:53 UTC] You have attempted to register a duplicate item with WooCommerce Navigation: `test-duplicate-item`
.....................................S....................... 2013 / 3898 ( 51%)
```

This simply changes the `error_log` call to `wc_doing_it_wrong`, which
uses WP Core's `_doing_it_wrong` function, which makes it so that the
error only gets output when WP_DEBUG is enabled. This should prevent
this message from being output in the middle of a unit test run.