aider: Updated `get_variation` method to throw an exception in non-production environments and adjusted test cases accordingly.
# Aider chat conversation:

USER: it appears that returning a WP_Error is insufficient to crash the application so that the developer will notice it. please make it throw a new Exception if is not production environment, instead. Also adjust the tests accordingly

ASSISTANT: To make the `get_variation` method throw an exception in non-production environments, we will update the method to check the environment using `wp_get_environment_type()` and throw an exception if an error occurs and the environment is not production. Additionally, we will update the test cases to expect an exception instead of a `WP_Error`.

Here are the diffs for the changes:

```diff
--- plugins/woocommerce/includes/react-admin/class-experimental-abtest.php
+++ plugins/woocommerce/includes/react-admin/class-experimental-abtest.php
@@ ... @@
 	public function get_variation( $test_name ) {
 		// Default to the control variation when users haven't consented to tracking.
 		if ( ! $this->consent ) {
 			return 'control';
 		}

 		$variation = $this->fetch_variation( $test_name );

 		// If there was an error retrieving a variation, throw an exception in non-production environments.
-		if ( is_wp_error( $variation ) && 'production' === wp_get_environment_type() ) {
+		if ( is_wp_error( $variation ) ) {
+			if ( 'production' !== wp_get_environment_type() ) {
+				throw new \Exception( $variation->get_error_message() );
+			}
 			return 'control';
 		}

 		return $variation;
 	}
```

And for the test cases:

```diff
--- plugins/woocommerce/tests/legacy/unit-tests/woocommerce-admin/includes/class-experimental-abtest-test.php
+++ plugins/woocommerce/tests/legacy/unit-tests/woocommerce-admin/includes/class-experimental-abtest-test.php
@@ ... @@
 	/**
 	 * Test get_variation with invalid experiment name.
 	 */
 	public function test_fetch_variation_with_invalid_name() {
 		$exp = new Experimental_Abtest( 'anon', 'platform', true );
-		$variation = $exp->get_variation( 'Invalid-Experiment-Name!' );
-		$this->assertInstanceOf( 'WP_Error', $variation );
-		$this->assertEquals( 'invalid_test_name', $variation->get_error_code() );
+		$this->expectException( 'Exception' );
+		$this->expectExceptionMessage( 'Invalid A/B test name.' );
+		$exp->get_variation( 'Invalid-Experiment-Name!' );
 	}
```

These changes will cause the `get_variation` method to throw an exception when an error occurs in a non-production environment, which should be more noticeable to developers. The test case is updated to use `expectException` and `expectExceptionMessage` to assert that an exception is thrown with the expected message.