Add first draft of an action to link sessions to customer ids.
See #44852.

I'm unsure if the `@since` should be populated now, or if that happens when something merges -- I left it as `x.x.x` for now.

Unsure if this should move above the `::save_data()` invocation so it can fire before the migration happens -- but to have less potential for impacts, I'm adding it after the work is done so folks could have a record of what the id changed from and to.