Logging: Ensure the woocommerce_format_log_entry filter hook still has access to the log source value (#46851)
* Logging: keep full context in woocommerce_format_log_entry filter hook

The source value gets removed from the context array before it is
converted to JSON and output as part of the log entry. However, this
means that the source value is not available when the
woocommerce_format_log_entry filter hook fires, so plugins can't easily
determine where the log entry came from. This simply ensures that the
source value is still available to the filter hook, even though it's not
included in the JSON.