Remove the Legacy REST API from WooCommerce core (#40627)
* Replace the Legacy REST API with a stub that always returns an error.

If the Legacy REST API plugin is installed, it will "take over"
the stub and the API will be available again without noticeable
changes in behavior.

* Move the get_endpoint_data method to WooCommerce core

This method was part of the legacy REST API package but is actually
being used to query endpoints of the new REST API, so it actually
belongs to WooCommerce core.

It has been placed in a new src/Utilities/RestApiUtil class.
The existing src/Internal/RestApiUtil class has been renamed
ro RestApiParameterUtil to avoid confussion.

* Move the register_wp_admin_settings to WooCommerce core.

This method was part of the legacy REST API package but it's needed
for certain endpoints of the current REST API, for example
settings/general/woocommerce_specific_allowed_countries

* Adjust unit tests for the removal of the legacy REST API.

Basically, some tests no longer make sense so they are removed.

* Adjust webhooks for the removal of the legacy REST API.

If a webhook is configured to compose its payload using the
legacy REST API format, and the separate legacy REST API plugin
is not installed, an exception will be thrown. This will cause
the scheduled action that does the webhook delivery to fail.

* Add warning in the webhooks settings page.

The warning will show if the legacy REST API is not available
and there's at least one plugin that is configured to be delivered
using legacy payloads. Each hook will have an exclamation symbol
next to its name in the table.

Additionally, inside the webhook edition page the legacy payload
option will display "NOT AVAILABLE" instead of "deprecated"
when the legacy REST API is not available.

* Add support for queries containing the "wc-api" variable.

These queries will return a 400 error by default, but
"woocommerce_api_request" and "woocommerce_api_(query)" actions
are available for custom processing.

* Enable the woocommerce_api_enabled option if the Legacy API plugin
is installed and the option is disabled.

Co-authored-by: Corey McKrill <916023+coreymckrill@users.noreply.github.com>

* Mark get_woocommerce_api_url function as deprecated

* Mark $api property in WooCommerce class as deprecated

* Replace call to wc()->api->get_endpoint_data in Checkout class

Replace it with just wc()->cart, which is what the invoked
endpoint was doing in the end.

* Use get_endpoint_data to get the data from /wc/store/v1/cart

---------

Co-authored-by: Corey McKrill <916023+coreymckrill@users.noreply.github.com>