Strategy for deciding when "undecided" means "incompatible" changed
Now, instead of a setting affecting all the plugins/features,
a per-feature flag named "plugins_are_incompatible_by_default"
is used. Defaulting to false, when it's true for a given feature
it means that plugins not declaring compatibility with a the
feature are to be considered incompatible with the feature.
The flag is set for the HPOS feature only.

Also a new 'woocommerce_plugins_are_incompatible_with_feature_by_default'
filter is added, which allows to alter the flag for a given feature.